app/requirements.txt
openai-whisper
ffmpeg-python
argostranslate
TTS==0.22.0
transformers==4.35.2
torch==2.1.2+cpu
torchaudio==2.1.2+cpu